Code management and renaming
-Removed unnecessary comments
-Renamed useExerciseContext.js and useSleepContext.js
