+++ client/src/context/ExerciseContext.js
@@ -20,15 +20,6 @@ export const ExerciseContextProvider = ({ children }) => {
         exercises: null
     })
 
-    // useEffect(() => {
-    //     const user = JSON.parse(localStorage.getItem('user'));
-    //     if (user) {
-    //         dispatch({ type: 'LOGIN', payload: user });
-    //     }
-    // }, [])
-
-    // console.log('AuthContext state: ', state);
-
     return (
         <ExerciseContext.Provider value={{...state, dispatch}}>
             { children }
