+++ client/src/context/SleepContext.js
@@ -20,15 +20,6 @@ export const SleepContextProvider = ({ children }) => {
         sleeps: null
     })
 
-    // useEffect(() => {
-    //     const user = JSON.parse(localStorage.getItem('user'));
-    //     if (user) {
-    //         dispatch({ type: 'LOGIN', payload: user });
-    //     }
-    // }, [])
-
-    // console.log('AuthContext state: ', state);
-
     return (
         <SleepContext.Provider value={{...state, dispatch}}>
             { children }
